Support TLS in TCP connectionst
This depends on prometheus/common#31
