+++ tcp.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"bufio"
+	"crypto/tls"
 	"fmt"
 	"net"
 	"net/http"
@@ -11,18 +12,36 @@ import (
 	"github.com/prometheus/log"
 )
 
+func dialTCP(target string, module Module) (net.Conn, error) {
+	dialer := &net.Dialer{Timeout: module.Timeout}
+	if !module.TCP.TLS {
+		return dialer.Dial("tcp", target)
+	}
+	config, err := module.TCP.TLSConfig.GenerateConfig()
+	if err != nil {
+		return nil, err
+	}
+	return tls.DialWithDialer(dialer, "tcp", target, config)
+}
+
 func probeTCP(target string, w http.ResponseWriter, module Module) bool {
 	deadline := time.Now().Add(module.Timeout)
-	conn, err := net.DialTimeout("tcp", target, module.Timeout)
+	conn, err := dialTCP(target, module)
 	if err != nil {
 		return false
 	}
 	defer conn.Close()
+
 	// Set a deadline to prevent the following code from blocking forever.
 	// If a deadline cannot be set, better fail the probe by returning an error
 	// now rather than blocking forever.
 	if err := conn.SetDeadline(deadline); err != nil {
 		return false
+	}
+	if module.TCP.TLS {
+		state := conn.(*tls.Conn).ConnectionState()
+		fmt.Fprintf(w, "probe_ssl_earliest_cert_expiry %f\n",
+			float64(getEarliestCertExpiry(&state).UnixNano())/1e9)
 	}
 	scanner := bufio.NewScanner(conn)
 	for _, qr := range module.TCP.QueryResponse {
